refactor: split final oversized rust files
